third_party/WebKit/LayoutTests/fast/preloader/style-base.html

<body>
<script>
if (window.testRunner)
    testRunner.dumpAsText();

window.addEventListener('DOMContentLoaded', function() {
    if (window.internals) {
        var success = internals.isPreloaded("resources/style1.css");
        console.log(success ? "PASS" : "FAIL");
    }
}, false);
</script>
This test requires content_shell --run-web-tests to see the log of what resources are loaded.
<script src=non-existant.js></script>
<script>document.write("<plaintext>");</script>
<base href="resources/foo.html">
<style>
@charset "ascii";
/* */
@import "style1.css";
</style>
