third_party/WebKit/LayoutTests/fast/preloader/style-base-expected.txt

CONSOLE MESSAGE: line 9: PASS
This test requires content_shell --run-web-tests to see the log of what resources are loaded.

<base href="resources/foo.html">
<style>
@charset "ascii";
/* */
@import "style1.css";
</style>
